Move Github request validation to github package
