+++ server/events/vcs/github/request_validator.go
@@ -11,7 +11,7 @@
 // limitations under the License.
 // Modified hereafter by contributors to runatlantis/atlantis.
 
-package server
+package github
 
 import (
 	"errors"
