+++ server/events/vcs/github/request_validator_test.go
@@ -11,7 +11,7 @@
 // limitations under the License.
 // Modified hereafter by contributors to runatlantis/atlantis.
 
-package server_test
+package github_test
 
 import (
 	"bytes"
@@ -20,14 +20,14 @@ import (
 	"testing"
 
 	. "github.com/petergtz/pegomock"
-	"github.com/runatlantis/atlantis/server"
+	"github.com/runatlantis/atlantis/server/events/vcs/github"
 	. "github.com/runatlantis/atlantis/testing"
 )
 
 func TestValidate_WithSecretErr(t *testing.T) {
 	t.Log("if the request is not valid against the secret there is an error")
 	RegisterMockTestingT(t)
-	g := server.DefaultGithubRequestValidator{}
+	g := github.DefaultGithubRequestValidator{}
 	buf := bytes.NewBufferString("")
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
@@ -42,7 +42,7 @@ func TestValidate_WithSecretErr(t *testing.T) {
 func TestValidate_WithSecret(t *testing.T) {
 	t.Log("if the request is valid against the secret the payload is returned")
 	RegisterMockTestingT(t)
-	g := server.DefaultGithubRequestValidator{}
+	g := github.DefaultGithubRequestValidator{}
 	buf := bytes.NewBufferString(`{"yo":true}`)
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
@@ -57,7 +57,7 @@ func TestValidate_WithSecret(t *testing.T) {
 func TestValidate_WithoutSecretInvalidContentType(t *testing.T) {
 	t.Log("if the request has an invalid content type an error is returned")
 	RegisterMockTestingT(t)
-	g := server.DefaultGithubRequestValidator{}
+	g := github.DefaultGithubRequestValidator{}
 	buf := bytes.NewBufferString("")
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
@@ -71,7 +71,7 @@ func TestValidate_WithoutSecretInvalidContentType(t *testing.T) {
 func TestValidate_WithoutSecretJSON(t *testing.T) {
 	t.Log("if the request is JSON the body is returned")
 	RegisterMockTestingT(t)
-	g := server.DefaultGithubRequestValidator{}
+	g := github.DefaultGithubRequestValidator{}
 	buf := bytes.NewBufferString(`{"yo":true}`)
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
@@ -85,7 +85,7 @@ func TestValidate_WithoutSecretJSON(t *testing.T) {
 func TestValidate_WithoutSecretFormNoPayload(t *testing.T) {
 	t.Log("if the request is form encoded and does not contain a payload param an error is returned")
 	RegisterMockTestingT(t)
-	g := server.DefaultGithubRequestValidator{}
+	g := github.DefaultGithubRequestValidator{}
 	buf := bytes.NewBufferString("")
 	req, err := http.NewRequest("POST", "http://localhost/event", buf)
 	Ok(t, err)
@@ -99,7 +99,7 @@ func TestValidate_WithoutSecretFormNoPayload(t *testing.T) {
 func TestValidate_WithoutSecretForm(t *testing.T) {
 	t.Log("if the request is form encoded and does not contain a payload param an error is returned")
 	RegisterMockTestingT(t)
-	g := server.DefaultGithubRequestValidator{}
+	g := github.DefaultGithubRequestValidator{}
 	form := url.Values{}
 	form.Set("payload", `{"yo":true}`)
 	buf := bytes.NewBufferString(form.Encode())
